rel method
